requirements.txt
google-generativeai
langchain
langchain-community
langchain-google-genai
sentence-transformers
faiss-cpu
pypdf
python-docx
python-dotenv
streamlit
